Move helpers out of physical plan optimizers
